Isolate popup logic, trivial refactors
